feat: added modifiable dtype for saving y file out

+++ hand_recognition/data_utils.py
@@ -36,7 +36,7 @@ def load_handtypes(dir_path):
                 data[folder][name] = torch.Tensor(np.load(os.path.join(dir_path,folder,f)))
     return data
 
-def save_all(trainX,trainY,valX,valY,parent_dir):
+def save_all(trainX,trainY,valX,valY,parent_dir,y_dtype='uint8'):
     """
     saves train,test folder numpy data from parent dir
     """
@@ -47,9 +47,9 @@ def save_all(trainX,trainY,valX,valY,parent_dir):
     if not os.path.isdir(test_dir):
         os.makedirs(test_dir)
     np.save(f"{os.path.join(parent_dir,'train')}/trainX",np.array(trainX).astype('uint8'))
-    np.save(f"{os.path.join(parent_dir,'train')}/trainY",np.array(trainY).astype('uint8'))
+    np.save(f"{os.path.join(parent_dir,'train')}/trainY",np.array(trainY).astype(y_dtype))
     np.save(f"{os.path.join(parent_dir,'val')}/valX",np.array(valX).astype('uint8'))
-    np.save(f"{os.path.join(parent_dir,'val')}/valY",np.array(valY).astype('uint8'))
+    np.save(f"{os.path.join(parent_dir,'val')}/valY",np.array(valY).astype(y_dtype))
 
 def save_data(data,path):
     directory = os.path.dirname(path)
